# Faster and more correct serialization