--- a/include/simdjson/common_defs.h
+++ b/include/simdjson/common_defs.h
@@ -6,8 +6,14 @@
 
 namespace simdjson {
 
-// defined in src/to_chars
+namespace internal {
+/**
+ * @private
+ * Our own implementation of the C++17 to_chars function.
+ * Defined in src/to_chars
+ */
 char *to_chars(char *first, const char *last, double value);
+}
 
 #ifndef SIMDJSON_EXCEPTIONS
 #if __cpp_exceptions