--- a/include/simdjson/dom/array.h
+++ b/include/simdjson/dom/array.h
@@ -9,7 +9,8 @@
 
 namespace simdjson {
 namespace dom {
-
+template<typename T>
+class string_builder;
 class document;
 class element;
 
@@ -126,6 +127,8 @@ private:
   friend struct simdjson_result<element>;
   template<typename T>
   friend class simdjson::minifier;
+  template<typename T>
+  friend class simdjson::dom::string_builder;
 };
 
 /**