--- a/include/simdjson/dom/object.h
+++ b/include/simdjson/dom/object.h
@@ -6,6 +6,10 @@
 #include "simdjson/internal/tape_ref.h"
 
 namespace simdjson {
+namespace internal {
+template<typename T>
+class string_builder;
+}
 namespace dom {
 
 class document;
@@ -209,7 +213,7 @@ private:
   friend class element;
   friend struct simdjson_result<element>;
   template<typename T>
-  friend class simdjson::dom::string_builder;
+  friend class simdjson::internal::string_builder;
 };
 
 /**