… benchmark/bench_dom_api.cpp
… include/simdjson/minify.h
… src/simdjson.cpp
… src/to_chars.cpp
… tests/CMakeLists.txt
… tests/basictests.cpp
… tests/minify_tests.cpp
… tests/test_macros.h