Currently, in simdjson, we have the recently introduced `std::minify` which can convert a JSON element back to a minified strings. Internally, it works with C++ streams. There are three problems with C++ streams...

1. It is slow. C++ streams are surprisingly slow.
2. It is locale-sensitive. 
3. It is lossy. Outputting a float to a stream does not at all guarantee that it can be later reparsed exactly.

This PR fixes these issues with `simdjson::minify`. It adds a new equivalent function `to_string`, as it is a standard way to convert objects to strings in C++11. I would also add `to_chars` to be C++17-ish, but I stopped short of it because C++17 is not our primary concern. 

This PR is backward compatible: there should be no API-breaking changes. However, the serialized output will differ. But that is fine. The new serialized output should be more precise and be locale independent.

It almost identical code from the previous code, I just copied @jkeiser 's code in a stream-free manner. The result is a 10x performance boost (see numbers below).  This is still very far from having optimal speed but it is probably good enough to eliminate serialization as a bottleneck. 

As stated elsewhere, this should be in 0.6 as I think it is of some importance.

```
$ ./benchmark/bench_dom_api --benchmark_filter=serialize
2020-09-12 00:22:12
Running ./benchmark/bench_dom_api
Run on (4 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x4)
Load Average: 0.24, 0.08, 0.05
-------------------------------------------------------------------------------------------------------------
Benchmark                                                   Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------------------
serialize_twitter/repeats:10_mean                     9945291 ns      9903803 ns           10 Gigabytes=47.1487M/s docs=100.981/s
serialize_twitter/repeats:10_median                   9910065 ns      9859544 ns           10 Gigabytes=47.3558M/s docs=101.425/s
serialize_twitter/repeats:10_stddev                    104686 ns       104290 ns           10 Gigabytes=486.462k/s docs=1.04188/s
serialize_twitter/repeats:10_max                     10212314 ns     10180782 ns           10 Gigabytes=47.5373M/s docs=101.813/s
serialize_twitter_to_string/repeats:10_mean            956152 ns       953911 ns           10 Gigabytes=490.768M/s docs=1048.33/s
serialize_twitter_to_string/repeats:10_median          956150 ns       953941 ns           10 Gigabytes=490.749M/s docs=1048.28/s
serialize_twitter_to_string/repeats:10_stddev            3272 ns         3041 ns           10 Gigabytes=1.56657M/s docs=3.34634/s
serialize_twitter_to_string/repeats:10_max             961239 ns       958123 ns           10 Gigabytes=493.872M/s docs=1054.96/s
serialize_twitter_string_builder/repeats:10_mean       868574 ns       866894 ns           10 Gigabytes=540.028M/s docs=1.15355k/s
serialize_twitter_string_builder/repeats:10_median     869201 ns       866731 ns           10 Gigabytes=540.127M/s docs=1.15376k/s
serialize_twitter_string_builder/repeats:10_stddev       2921 ns         1887 ns           10 Gigabytes=1.17572M/s docs=2.51144/s
serialize_twitter_string_builder/repeats:10_max        872932 ns       869419 ns           10 Gigabytes=541.723M/s docs=1.15717k/s
```

There are other benefits to this approach as it opens the door to the addition of a prettifier.


Fixes https://github.com/simdjson/simdjson/issues/933

According to @plokhotnyuk The floating-point serialization could probably be even better, but this can be handled in the future.