@plokhotnyuk 2020-09-12
@lemire Have you considered using the amazing work of Raffaello Giulietti ["The Schubfach way to render doubles"](https://drive.google.com/file/d/1luHhyQF9zKlM8yJ1nebU0OgVYhfC6CBN/view) to serialize 64-bit floating point numbers into shortest and the most precise text representation?

[Here](https://github.com/plokhotnyuk/jsoniter-scala/blob/14b3b21e2f00045621fab9bda215c3e4d5742931/jsoniter-scala-core/jvm/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonWriter.scala#L1839-L1891) is an adoption of the Schubfach way in one of the fastest JSON serializer for Scala.