@lemire 2020-09-12
@plokhotnyuk Do you have a link to a C/C++ along with benchmarks?



At this point in time, I am not even trying to achieve best performance, let alone optimal. I am moving us off the C++ streams. Period.

If there is ready and tested code that I could adopt, please share... but it needs to be mature code.