@jkeiser 2020-09-12
I am still flabbergasted that C++ streams are this slow. I figured it'd be 50% worse, or something, and I could live with that at the time, but wow. Just wow.