@lemire 2020-09-12
> I am still flabbergasted that C++ streams are this slow.

I hear that they are the root cause of the fires on the West Coast.