@lemire 2020-09-12
@plokhotnyuk  I have looked at the C++ implementations and none of them look mature enough. It may well be great work, but I don't trust the look of the code, and I do not want to spend the time at this point to examine it.