@lemire 2020-09-13
@jkeiser  This is ready for review.This modifies your code, so please review.

It can be improved. At this point, it is enough that it is correct, and much faster.