@lemire 2020-09-14
I'll take your concerns point by point and try to address them. It is important to get these things right.