@lemire 2020-09-14
> OK, this is just a thought. Don't take it as a request for change, just mull it over.

This might be totally fine but I don't want to mix API design with this PR. I am mostly preoccupied with getting the fundamentals right. So what I'll do is hide away everything in "internal" and make it private, so that we have the freedom of tweaking the API later.