@lemire 2020-09-14
@jkeiser My latest commit hides away string_buffer and mini_formatter in the internal namespace and I have put clear warning that it is not part of our public API.

So we can take a second pass and expose whatever we like later.