@jkeiser 2020-09-14
> > This gives you the option to keep the formatter around between invocations, which I think is one of
> > the reasons you wrote string_builder :)
> 
> My thinking was as follows. Suppose that you have an application where you routine grab a subset of a JSON document and you pass it along. The way we do things currently, we would constantly be allocating memory to hold the string content _even_ if all of your strings are about the same size. In such a scenario, you'd want to keep reusing your buffer. But my thinking is not very relevant since I don't have a real use case for it. So let us hide all this where the users won't see it. This way, we can change it later.

It's a good thought, and I agree anytime we have allocation, we should let people reuse it.