@lemire 2020-09-23
Merging.